Design Documents
I created a new folder for our design documents as well as uploaded the
ones that I have completed for now. The Puzzle design document is
currently lacking information for the latter portions due to it being
not fully discussed as of yet. After I get some more information, I will
go back and fill them in.